added cast to clucene's Hits object uncommented genID
git-svn-id: https://crosswire.org/svn/sword/trunk@2670 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ src/modules/swmodule.cpp
@@ -515,7 +515,7 @@ ListKey &SWModule::search(const char *istr, int searchType, int flags, SWKey *sc
 
 			// iterate thru each good module position that meets the search
 			bool checkBounds = getKey()->isBoundSet();
-			for (unsigned long i = 0; i < h->length(); i++) {
+			for (unsigned long i = 0; i < (unsigned long)h->length(); i++) {
 				Document &doc = h->doc(i);
 
 				// set a temporary verse key to this module position
